requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-cors-headers==3.4.0
uWSGI>=2.0.18,<2.1